+++ backEndDataStorage/SpringWebServer/src/main/java/com/edg/Dao/ServerDaoImpl.java
@@ -70,13 +70,13 @@ public class ServerDaoImpl implements ServerDao {
         return getDataFromDB(sql);
     }
 
-    public void saveTaskInDB(String sqlQuery, String taskName, String taskBody, boolean isCompleted, int timestamp) {
+    public void saveTaskInDB(String sqlQuery, String taskName, String taskBody, boolean isCompleted, long timestamp) {
         try (Connection conn = this.connect();
              PreparedStatement preparedStatement = conn.prepareStatement(sqlQuery)) {
             preparedStatement.setString(1, taskName);
             preparedStatement.setString(2, taskBody);
             preparedStatement.setBoolean(3, isCompleted);
-            preparedStatement.setInt(4, timestamp);
+            preparedStatement.setLong(4, timestamp);
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
             System.out.println(e.getMessage());
